requirements.txt
tweepy
pandas
apache-airflow
